ADD: fixed int to size_t when overload during upload

+++ inc/Config/ConfigUtils.hpp
@@ -6,7 +6,7 @@ struct ServerConfig;
 struct Directory;
 
 std::string removeSpace( std::string &line );
-int getMaxRequestBody(std::string &noSpaceLine);
+long long getMaxRequestBody(std::string &noSpaceLine);
 std::string getInfo( std::string &noSpaceLine );
 Directory &findPath( ServerConfig server, std::string path );
 bool stringToSizeT(const std::string& str, size_t& result);
